chore: rename to prepare and move prettyprint over
